requirements.txt
psycopg2-binary
Flask-SQLAlchemy
Flask-Script
Flask-Migrate
requests
